Cache if buffer is not in a project
IF a buffer is not in a project, running project-current in the mode
line can become very expensive because it has to constantly retraverse
the file’s tree. This is not a problem for files in a project because
once their project is found, the result is cached using
`vc-file-setprop'. So, it makes sense to remember whether this buffer
is in a project or not.